Move tests to previous render body migration test
